Compute max generated block size depending on UAHF activation
Summary: We don't want to allow block more than 1MB to be generated before UAHF activation.

Test Plan: Added tests that check the generated block size before and after activation.

Reviewers: freetrader, sickpig, kyuupichan, awemany, #bitcoin_abc

Reviewed By: freetrader, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D206

+++ src/miner.h
@@ -135,7 +135,7 @@ private:
     CBlock *pblock;
 
     // Configuration parameters for the block size
-    unsigned int nMaxGeneratedBlockSize;
+    uint64_t nMaxGeneratedBlockSize;
     CFeeRate blockMinFeeRate;
 
     // Information on the current status of the block
@@ -163,9 +163,7 @@ public:
     std::unique_ptr<CBlockTemplate>
     CreateNewBlock(const CScript &scriptPubKeyIn);
 
-    unsigned int GetMaxGeneratedBlockSize() const {
-        return nMaxGeneratedBlockSize;
-    }
+    uint64_t GetMaxGeneratedBlockSize() const { return nMaxGeneratedBlockSize; }
 
 private:
     // utility functions
